Move meta methods from package controller to a new controller
