+++ src/api/app/views/webui/package/_tabs.html.haml
@@ -11,4 +11,4 @@
       = tab_link('Requests', package_requests_path(project, package), false, 'scrollable-tab-link')
       = tab_link('Users', package_users_path(project, package), false, 'scrollable-tab-link')
       = tab_link('Attributes', index_attribs_path(project, package), false, 'scrollable-tab-link')
-      = tab_link('Meta', package_meta_path(project, package), false, 'scrollable-tab-link')
+      = tab_link('Meta', project_package_meta_path(project, package), false, 'scrollable-tab-link')
