+++ src/api/app/views/webui/packages/meta/_breadcrumb_items.html.haml
@@ -1,0 +1,1 @@
+= render partial: 'webui/package/breadcrumb_items'
