+++ src/api/spec/controllers/webui/package_controller_spec.rb
@@ -40,19 +40,6 @@ RSpec.describe Webui::PackageController, :vcr do
     HEREDOC
   end
 
-  describe 'GET #meta' do
-    before do
-      get :meta, params: { project: source_project, package: source_package }
-    end
-
-    it 'sends the xml representation of a package' do
-      expect(assigns(:meta)).to eq(source_package.render_xml)
-    end
-
-    it { expect(response).to render_template('package/meta') }
-    it { expect(response).to have_http_status(:success) }
-  end
-
   describe 'POST #remove' do
     before do
       login(user)
@@ -301,106 +288,6 @@ RSpec.describe Webui::PackageController, :vcr do
       it { expect(a_request(:post, post_url)).not_to have_been_made }
       it { expect(flash[:error]).to eq('Sorry, you are not authorized to update this package.') }
       it { is_expected.to redirect_to(root_path) }
-    end
-  end
-
-  describe 'POST #save_meta' do
-    let(:valid_meta) do
-      "<package name=\"#{source_package.name}\" project=\"#{source_project.name}\">" \
-        '<title>My Test package Updated via Webui</title><description/></package>'
-    end
-
-    let(:invalid_meta_because_package_name) do
-      "<package name=\"whatever\" project=\"#{source_project.name}\">" \
-        '<title>Invalid meta PACKAGE NAME</title><description/></package>'
-    end
-
-    let(:invalid_meta_because_project_name) do
-      "<package name=\"#{source_package.name}\" project=\"whatever\">" \
-        '<title>Invalid meta PROJECT NAME</title><description/></package>'
-    end
-
-    let(:invalid_meta_because_xml) do
-      "<package name=\"#{source_package.name}\" project=\"#{source_project.name}\">" \
-        '<title>Invalid meta WRONG XML</title><description/></paaaaackage>'
-    end
-
-    before do
-      login user
-    end
-
-    context 'with proper params' do
-      before do
-        post :save_meta, params: { project: source_project, package: source_package, meta: valid_meta }
-      end
-
-      it { expect(flash[:success]).to eq('The Meta file has been successfully saved.') }
-      it { expect(response).to have_http_status(:ok) }
-    end
-
-    context 'without admin rights to raise protection level' do
-      before do
-        allow_any_instance_of(Package).to receive(:disabled_for?).with('sourceaccess', nil, nil).and_return(false)
-        allow(FlagHelper).to receive(:xml_disabled_for?).with(Xmlhash.parse(valid_meta), 'sourceaccess').and_return(true)
-
-        post :save_meta, params: { project: source_project, package: source_package, meta: valid_meta }
-      end
-
-      it { expect(flash[:error]).to eq('Error while saving the Meta file: admin rights are required to raise the protection level of a package.') }
-      it { expect(response).to have_http_status(:bad_request) }
-    end
-
-    context 'with an invalid package name' do
-      before do
-        post :save_meta, params: { project: source_project, package: source_package, meta: invalid_meta_because_package_name }
-      end
-
-      it { expect(flash[:error]).to eq('Error while saving the Meta file: package name in xml data does not match resource path component.') }
-      it { expect(response).to have_http_status(:bad_request) }
-    end
-
-    context 'with an invalid project name' do
-      before do
-        post :save_meta, params: { project: source_project, package: source_package, meta: invalid_meta_because_project_name }
-      end
-
-      it { expect(flash[:error]).to eq('Error while saving the Meta file: project name in xml data does not match resource path component.') }
-      it { expect(response).to have_http_status(:bad_request) }
-    end
-
-    context 'with invalid XML' do
-      before do
-        post :save_meta, params: { project: source_project, package: source_package, meta: invalid_meta_because_xml }
-      end
-
-      it do
-        expect(flash[:error]).to match(/Error while saving the Meta file: package validation error.*FATAL:/)
-        expect(flash[:error]).to match(/Opening and ending tag mismatch: package line 1 and paaaaackage\./)
-      end
-
-      it { expect(response).to have_http_status(:bad_request) }
-    end
-
-    context 'when connection with the backend fails' do
-      before do
-        allow_any_instance_of(Package).to receive(:update_from_xml).and_raise(Backend::Error, 'fake message')
-
-        post :save_meta, params: { project: source_project, package: source_package, meta: valid_meta }
-      end
-
-      it { expect(flash[:error]).to eq('Error while saving the Meta file: fake message.') }
-      it { expect(response).to have_http_status(:bad_request) }
-    end
-
-    context 'when not found the User or Group' do
-      before do
-        allow_any_instance_of(Package).to receive(:update_from_xml).and_raise(NotFoundError, 'fake message')
-
-        post :save_meta, params: { project: source_project, package: source_package, meta: valid_meta }
-      end
-
-      it { expect(flash[:error]).to eq('Error while saving the Meta file: fake message.') }
-      it { expect(response).to have_http_status(:bad_request) }
     end
   end
